refactor: move logic into src directory
